Changed the entity access level in the DbContext
The entity was public earlier, thus allowing direct access even when the
methods to operate on entity was defined in the context. Now the entity
is private and the only way to manipulate data is using the methods in
the DbContext thus ensuring, consumers use the context and not the
entity.

Defined the constant for RELEASE mode. This allows the DbContext to be
performant in release mode by shedding away the virtual method
declaration to concrete methods.

Added the binary column to store assets like images / videos