Worked on Issue #3
There were tons of challenges. The code being committed is only
relatively stable.
The primary problems were -
1. If pings are only from one pinger then the target pinger never gets
to see the source pinger in Ping List
In the process of fixing this defect, I faced the challenge of
displaying the right profile picture in ping history list.
2. The last posted message is always the last message of the current
pinger.. However, it should be the last message from the entire ping
exchanges.. :disappointed: